==> 1a. DATA PREPROCESSING - DATA MERGE.ipynb <==
# 1: DATA PREPROCESSING

The objective of this step is to merge datasets to create a comprehensive dataset about movie titles. This includes information on basic title details, financial performance, viewer ratings, and cast and crew.



import pandas as pd
import numpy as np

## BASICS

The `title.basics.tsv.gz` dataset from IMDb contains basic information about titles listed in the IMDb database. This file is essential for getting an overview of each title's fundamental attributes. The movies are filtered for the scope of this study.

basics_df = pd.read_csv('./raw_data/title.basics.tsv.gz', compression='gzip', delimiter='\t')
movies_df = basics_df.loc[basics_df['titleType']=='movie'].copy()
movies_df.head()

## RATINGS

The `ratings.tsv.gz` dataset from IMDb contains viewer ratings for titles in the IMDb database. It is used to assess the popularity and viewer reception of various titles, and trends in audience preferences and title performance.

ratings_df = pd.read_csv('./raw_data/title.ratings.tsv.gz',compression ='gzip',delimiter='\t')
ratings_df.head()

## PRINCIPALS

The `principals.tsv.gz` dataset from IMDb details the principal cast and crew of the titles.This dataset is crucial for understanding who played key roles in the production and performance of a title.

Preprocessing was done in order to list all names under a single column.

principals_df = pd.read_csv('./raw_data/title.principals.tsv.gz',
                            compression ='gzip',delimiter='\t')
principals_restructured = principals_df.groupby(['tconst', 'category'],
                                                as_index=False)['nconst'].agg(lambda x: ','.join(x))
principals_pivoted = principals_restructured.pivot(index='tconst',
                                                   columns='category',
                                                   values='nconst').reset_index()

# FOR VIEWING OF NON-NULL COLUMNS ONLY
principals_pivoted.loc[(principals_pivoted.isnull().sum(axis=1)
                        .sort_values().head().index)]

## GROSS INCOME

This dataset is sourced from Kaggle ([link](https://www.kaggle.com/datasets/ashishjangra27/imdb-movies-dataset)). It provides financial information about the titles in the IMDb database. This was used as additional information analyze the financial success of movies and other titles, offering insights into the movies' market performance.

Cleaning was done to remove 1000 separators, currency units, and convert this data into `float`.

income_df = pd.read_csv('./raw_data/movies.csv', usecols=['id','gross_income'])
income_df.columns = ['tconst', 'gross_income']

# CLEAN INCOME DATA
income_df['gross_income'] = income_df['gross_income'].str.replace(',','')
income_df['gross_income'] = income_df['gross_income'].str.replace('$','')
income_df['gross_income'] = income_df['gross_income'].str.replace('M','')
income_df['gross_income'] = income_df['gross_income'].astype(float)

income_df.sort_values('gross_income', ascending=False).head()

## MERGE DATASETS

The merging all of these datasets on the `tconst` column aims to create a single  dataset about movie titles in the IMDb database. By using Pandas' merge function in Python, these datasets are combined based on the `tconst `identifier, joining essential information such as basic title details, financials, ratings, and cast and crew data.

This unified dataset will be used for deeper analysis of movie titles.

# Merge basics_df and ratings_df on the 'tconst' column
merged_df1 = pd.merge(movies_df, ratings_df, on='tconst', how='inner')
print(merged_df1.shape)

# Merge the resulting DataFrame with principals_df on the 'tconst' column
merged_df2 = pd.merge(merged_df1, principals_pivoted, on='tconst', how='left')
print(merged_df2.shape)

# Merge the resulting DataFrame with principals_df on the 'tconst' column
merged_df3 = pd.merge(merged_df2, income_df, on='tconst', how='left').set_index('tconst')
print(merged_df3.shape)

# remove irrelevant columns

remove_cols = ['titleType', 'originalTitle', 'endYear','archive_footage', 'archive_sound']

merged_df = merged_df3.drop(remove_cols, axis=1)
merged_df = merged_df.replace("\\N",np.nan)

# Display the merged DataFrame
merged_df = merged_df.dropna(subset='genres').copy()

merged_df['gross_income'] = merged_df['gross_income'].fillna(0)
merged_df.head()

## SAVE DATASET

The dataset is saved for use in the succeeding notebooks.

merged_df.to_csv('processed_data/data.csv')
