==> 2. FEATURE ENGINEERING.ipynb <==
# 2: FEATURE ENGINEERING

Derive a composite score that takes into account popularity, ratings, and profitability into account. Also, target encoding was used to convert categories into continuous values.


## STEP 2a: COMPOSITE SCORE

This outlines the method in defining a metric that captures the notion of a “hit movie”. In the merged dataset. The following features was used in producing a single metric:

- **NUMBER OF VOTES.** Servers as an indicator of the title's popularity and viewer engagement.

- **GROSS INCOME**. A key indicator of the title's financial success and market appeal.
- **AVERAGE RATINGS.**
Reflects the overall viewer reception and satisfaction. Higher ratings often indicate a more favorable response from the audience.

import pandas as pd
import numpy as np

# Load data
data = pd.read_csv('processed_data/data.csv', index_col=0)
data.head()

data['gross_income_log'] = np.log10(data['gross_income']+1)
data['numVotes_log'] = np.log10(data['numVotes']+1)

### THE CRITIC METHOD

The CRITIC (Criteria Importance Through Intercriteria Correlation) method, developed by Diakoulaki, Mavrotas, and Papayannakis in 1995, is a technique in Multi-Criteria Decision-Making (MCDM).
It assesses the importance of criteria based on their variance and the correlation between them. Criteria that provide unique information by being less correlated with others and showing higher variance are given higher weights, thus improving decision-making objectivity.

#### STEPS

1. Normalize (min-max normalization)  the decision matrix.
2. Calculate the standard deviation for each criteria on the normalized matrix ($\sigma_i$)
3. Determine the correlation matrix $r_{ij}$
4. Calculate the measure of conflict
$\displaystyle\sum^{n}_{j=1}r_{ij}$
5. Determine the quantity of information associated with each criterion:
$c_i=\sigma_i\displaystyle\sum^{n}_{j=1}r_{ij}$
6. Determine the weights $w_i=\frac{c_i}{\sum^{n}_{i=1}c_i}$

# Shuffle the indices of the DataFrame
shuffled_indices = np.random.permutation(len(data))

# Calculate the sizes for each split
train_size = int(0.6 * len(data))
val_size = int(0.2 * len(data))
test_size = len(data) - train_size - val_size  # This ensures that rounding issues don't leave out any data

# Split the data into three parts
data_train2 = data.iloc[shuffled_indices[:train_size]].copy()
data_train2 = data_train2[['numVotes_log','gross_income_log','averageRating']]

data_train2 = data_train2[['numVotes_log','gross_income_log','averageRating']]

from sklearn.preprocessing import MinMaxScaler

mm = MinMaxScaler().fit(data_train2)

normalized_data_train2 = pd.DataFrame(mm.transform(data_train2),
                                         columns = data_train2.columns,
                                         index = data_train2.index)

normalized_data_train2_nozero = normalized_data_train2.loc[normalized_data_train2['gross_income_log']>0]
conflict = 1 - normalized_data_train2_nozero.corr()
conflict['sum'] = conflict.sum(axis=1)
conflict['stdev'] = data_train2.std()
conflict['info_qty'] = conflict['sum']*conflict['stdev']
weights = conflict['info_qty']/conflict['info_qty'].sum()
weights

data2 = data[['numVotes_log','gross_income_log','averageRating']].copy()

normalized_data = pd.DataFrame(mm.transform(data2),
                               columns = data2.columns,
                               index = data2.index)

data['composite_score'] = (normalized_data*weights).sum(axis=1)
data['composite_score'].describe()

## STEP 2b: TARGET ENCODING

A technique for encoding categorical variables using the average value of the target variable. This method transforms categorical features into continuous values.

### ADVANTAGES
* Reduces overfitting compared to traditional one-hot encoding.
* Handles high cardinality categorical data effectively.

### PROCESS
1. Group data by categorical feature.
2. Calculate the mean of the target variable for each category.
3. Replace the categorical variable with the calculated mean.

data_TE = data[['primaryTitle','isAdult','startYear',
                          'averageRating','runtimeMinutes',
                          'numVotes_log','gross_income_log','composite_score']].copy()
data_TE.head()

fill_cols = ['genres','actor','actress','casting_director','cinematographer',
             'composer','director','editor','producer','production_designer',
             'self','writer']

cols_to_catmeans = {}

for col in fill_cols:

    data[col] = data[col].fillna('')
    data[f'{col}_temp'] = data[col].str.split(',')
    category_means = data[[f'{col}_temp','composite_score']].explode(f'{col}_temp').groupby(f'{col}_temp')['composite_score'].mean()
    
    cols_to_catmeans[col] = category_means
    # Step 3: Replace categories with corresponding mean target values
    data_TE[f'{col}_TE'] = data[f'{col}_temp'].apply(lambda x: np.mean([category_means[c] for c in x]))

    data_TE[f'{col}_COUNT'] = data[f'{col}_temp'].apply(len)

### SPLIT AND SAVE THE TRAIN, VALIDATION, AND TEST DATASETS

# Split the data into three parts
data_train = data_TE.iloc[shuffled_indices[:train_size]].copy().dropna()
data_val = data_TE.iloc[shuffled_indices[train_size:train_size + val_size]].copy().dropna()
data_test = data_TE.iloc[shuffled_indices[train_size + val_size:]].copy().copy().dropna()

data_test.head()

data_train.to_csv('processed_data/train.csv')
data_val.to_csv('processed_data/val.csv')
data_test.to_csv('processed_data/test.csv')

## SAVE THE TARGET ENCODINGS
The category means are saved to decode values into categories.

import pickle
from datetime import datetime

curr_time = datetime.today().strftime('%Y-%m-%d %H-%M-%S')

FILENAME = f'./models/{curr_time}_col_to_catmeans.pkl'

pickle.dump(cols_to_catmeans, open(FILENAME, 'wb'))
print(FILENAME)
