==> 5. TOPIC ANALYSIS ON TITLES SIMILAR TO GLOBAL MAXIMA.ipynb <==
# 5: TOPIC ANALYSIS ON TITLES SIMILAR TO GLOBAL MAXIMA

Analyze keywords of titles whose features are similar to the global maximum of the composite score model.

## Step 5a: Recap of Global Maxima


# Load the pickled file
import pickle
import pandas as pd
pd.set_option('display.max_rows', 500)


# MODEL_FILENAME = './models/2024-05-08 19-28-45_Booster.pkl'
OPTIMALFEATURES_FILENAME = './models/2024-05-13 12-43-07_OPTIMALFEATURES.pkl'

with open(OPTIMALFEATURES_FILENAME, 'rb') as file:
    best_params = pickle.load(file)
    
best_params

## Step 5b: Decode Target Encoded Values

We aim to find data points that are similar in feature values to the global maximum, based on Euclidean distance.

CATMEANS_FILENAME = './models/2024-05-11 18-59-46_col_to_catmeans.pkl'

with open(CATMEANS_FILENAME, 'rb') as file:
    col_to_catmeans = pickle.load(file)

# Function to find closest n values to x
def find_closest_values(series, x, n):
    # Calculate the absolute differences from x
    differences = series.sub(x).abs()
    # Sort these differences and get the top n closest
    closest_indices = differences.nsmallest(n).index
    return series.loc[closest_indices]

features = ['genres','actor','actress','casting_director','cinematographer',
             'composer','director','editor','producer','production_designer',
             'self','writer']

records = []
for feat in features:
    series = col_to_catmeans[feat.replace('_temp','')]
    x = best_params[f'{feat}_TE']
    n = best_params[f'{feat}_COUNT']
    
    for close_cat in find_closest_values(series, x, n).index:
        records.append([feat, close_cat])

Using `col_to_catmeans`, which stores the mean scores for each category, continuous values are decoded back to their respective categories.

Given that the values of the global minimum might not match exactly with the category means in `col_to_catmeans`, the closest target encoded value to the actual value will be selected.

**Example**

    'genres_TE': 0.3844082717006556,
    'genres_COUNT': 3
    
This conversion calls for decoding the value of `genres_TE` into the 3 closest target-encoded categories to the `genres_TE` value. Which means that the 3 closest categories are the following:
* Biography (`0.381871`)
* Sport (`0.363881`)
* History (`0.359789`)

col_to_catmeans['genres']

We do the same thing to decode the other feature values into categories.

import pandas as pd
crew = pd.DataFrame(records, columns=['feature', 'nconst'])

names = pd.read_csv('./raw_data/name.basics.tsv.gz', compression='gzip',
                    delimiter='\t', usecols=['nconst','primaryName'])

crew_with_names = pd.merge(crew, names, how='left', on='nconst')
crew_with_names['primaryName'] = crew_with_names['primaryName'].fillna(crew_with_names['nconst'])
crew_with_names = crew_with_names.drop('nconst',axis=1)

# Grouping by 'feature' and joining 'primaryName' values with commas
crew_with_names = crew_with_names.groupby('feature')['primaryName'].agg(', '.join).reset_index()
display(crew_with_names)

### Save dataset

crew_with_names.to_csv('./processed_data/ultimate_crossover.csv')

## Step 5c: Topic Analysis on Titles Similar to Global Maxima

## LOAD ENTIRE DATASET

# Load data
train = pd.read_csv('processed_data/train.csv', index_col=0).dropna()
val = pd.read_csv('processed_data/val.csv', index_col=0).dropna()
test = pd.read_csv('processed_data/test.csv', index_col=0).dropna()

data = pd.concat([train, val, test])
data_feat = data[list(best_params.keys())].copy()

Before getting the closest values, it is necessary to scale both the data and the global maximum first.

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
data_feat_mm = pd.DataFrame(scaler.fit_transform(data_feat),
                            columns=data_feat.columns,
                            index=data_feat.index)
data_feat_mm.head()

Once scaled, the closest 100 titles with feature values similar to the global maximum as are listed below.

best_params_df = pd.DataFrame(pd.Series(best_params)).T
best_params_df = pd.DataFrame(scaler.transform(best_params_df[data_feat_mm.columns]),
                              columns=best_params_df.columns)
best_params_scaled = best_params_df.loc[0]

import pandas as pd
from scipy.spatial import distance

# Assuming data is your DataFrame with the dataset

# Function to calculate Euclidean distance
def calculate_distance(row):
    return distance.euclidean(row, best_params_scaled)

# Calculate distance from each row in the DataFrame to the best_params
data_feat_dist = data_feat.apply(calculate_distance, axis=1)

# Sort by distance and select the top 100 closest entries
closest_100 = data_feat_dist.sort_values().head(100)
movie_titles = data.loc[closest_100.index,'primaryTitle']
movie_titles.head(100)

## Retrieving the most popular terms

TF-IDF, or Term Frequency-Inverse Document Frequency, is a statistical measure used to evaluate the importance of a word in a collection of documents. It helps identify the most relevant terms in each document.
* The **term frequency (TF)** part assesses how frequently a term appears in a specific document,
* while the **inverse document frequency (IDF)** part evaluates how rare or common the term is across the whole document set.
The combination of these two measures helps to prioritize terms that are frequent in a particular document but not too common across all documents, making them distinctive and potentially more relevant to the topic of the document. This makes TF-IDF particularly useful for extracting top terms that uniquely characterize each document.

from sklearn.feature_extraction.text import TfidfVectorizer

# Initialize the TF-IDF Vectorizer
vectorizer = TfidfVectorizer(ngram_range=(1,3))

# Fit and transform the series
tfidf_matrix = vectorizer.fit_transform(movie_titles)

# Get feature names to use as dataframe column headers
feature_names = vectorizer.get_feature_names_out()

# Create a DataFrame for the TF-IDF values
df_tfidf = pd.DataFrame(tfidf_matrix.toarray(), columns=feature_names)

Using this approach, the top 3 terms are the following:

# Function to get top n terms with highest tf-idf scores for each document
def get_top_n_terms(dataframe, top_n=5):
    top_terms = {}
    for index, row in dataframe.iterrows():
        # Sort the TF-IDF values in descending order and select the top n
        sorted_row = row.sort_values(ascending=False)
        top_terms[index] = sorted_row.head(top_n).index.tolist()
    return top_terms

top_terms = 20
# Get top 5 terms for each document
top_terms_per_document = get_top_n_terms(df_tfidf, top_terms)

# If you want to find the overall top terms in the corpus:
top_terms_overall = df_tfidf.max(axis=0).sort_values(ascending=False).head(top_terms).index.tolist()

print("Top overall terms:", top_terms_overall)

## Suggested Title and Synopses from ChatGPT

The following top terms, I asked ChatGPT to generate titles and synopses with the following terms:

|     Title                                   |     Synopsis                                                                                                                                               |
|---------------------------------------------|------------------------------------------------------------------------------------------------------------------------------------------------------------|
|     Transit   Through Time                  |     A   historical sports biography about an athlete overcoming adversities in   different eras.                                                           |
|     Extra   Presidential                    |     A   biographical film about a lesser-known but impactful president whose policies   left a lasting mark.                                               |
|     Prosopagnosia:   The Unseen Faces       |     A   deep, personal story about someone struggling with face blindness, possibly   an athlete or historical figure who must overcome this challenge.    |
|     Pupa:   Transformation of a Champion    |     A   biography focusing on an athlete's metamorphosis and trials.                                                                                       |
|     Unhinged   Melodies                     |     With   music by Arcade Fire, a film about an unstable but genius musician who   influenced a sport or historical event.                                |

## REFERENCES
* The CRITIC Method. *New Methods and Applications in Multiple Attribute Decision Making (MADM)*
* The Beauty of Bayesian Optimization, Explained in Simple Terms. https://towardsdatascience.com/the-beauty-of-bayesian-optimization-explained-in-simple-terms-81f3ee13b10f
* ChatGPT prompts to generate movie titles of top terms.
